+++ src/components/List.ts
@@ -162,6 +162,7 @@
         }
 
 
+
         public addChildAt(child, index) {
             var result = super.addChildAt(child, index);